Refactor plan usage counting from Integer to Long and optimize queries
- Change limit and currentUsage types from Integer/int to Long/long across all plan-related classes
- Optimize GenericRepository by adding count methods instead of fetching all records and getting size
- Update IPlanUsageService and PlanUsageServiceImpl to use direct count queries for better performance
- Add JavaDoc documentation and @RegisterForReflection annotation to APILimitExceptionDTO
- Update PlanException to handle Long types for limit tracking

+++ xpenbox-back/src/main/java/org/xpenbox/enforcement/service/IPlanUsageService.java
@@ -8,21 +8,21 @@ public interface IPlanUsageService {
     /**
      * Counts the number of user accounts, credit cards, and categories for a given user ID.
      * @param userId the ID of the user for whom the plan usage is to be counted
-     * @return an integer representing the count of user accounts, credit cards, and categories for the specified user
+     * @return a long representing the count of user accounts, credit cards, and categories for the specified user
      */
-    int countUserAccounts(Long userId);
+    long countUserAccounts(Long userId);
 
     /**
      * Counts the number of credit cards and categories for a given user ID.
      * @param userId the ID of the user for whom the plan usage is to be counted
-     * @return an integer representing the count of credit cards and categories for the specified user
+     * @return a long representing the count of credit cards and categories for the specified user
      */
-    int countUserCreditCards(Long userId);
+    long countUserCreditCards(Long userId);
 
     /**
      * Counts the number of categories for a given user ID.
      * @param userId the ID of the user for whom the plan usage is to be counted
-     * @return an integer representing the count of categories for the specified user
+     * @return a long representing the count of categories for the specified user
      */
-    int countUserCategories(Long userId);
+    long countUserCategories(Long userId);
 }
